Bounds/poly zoom: more responsive; break out point hooks
