+++ src/components/map/GeocoderPopout.tsx
@@ -7,9 +7,8 @@ import { createStyles, makeStyles, Theme } from '@material-ui/core'
 import 'react-map-gl-geocoder/dist/mapbox-gl-geocoder.css'
 
 import { useMapToolsDispatch } from 'components/context'
-import { useUItext } from 'components/generic'
+import { useBreakpoint, useUItext } from 'components/generic'
 import { MAPBOX_TOKEN, NYC_LAT_LONG, POINT_ZOOM_LEVEL } from './config'
-import { useWindowResize } from '../../utils'
 import { flyToBounds, flyToPoint } from './utils'
 import { GeocodeResult, BoundsArray, GeocoderPopoutProps } from './types'
 
@@ -52,7 +51,7 @@ const useStyles = makeStyles((theme: Theme) =>
 export const GeocoderPopout: FC<GeocoderPopoutProps> = (props) => {
   const { mapRef } = props
   const geocoderContainerRef = useRef<HTMLDivElement>(null)
-  const { width, height } = useWindowResize()
+  const breakpoint = useBreakpoint()
   const { text: placeholderText } = useUItext('loc-search-placeholder')
   const classes = useStyles()
   const mapToolsDispatch = useMapToolsDispatch()
@@ -70,13 +69,13 @@ export const GeocoderPopout: FC<GeocoderPopoutProps> = (props) => {
     if (bbox) {
       // TODO: if (geocodeResult.result.place_type[0] === 'neighborhood')...
       const settings = {
-        height,
-        width,
+        height: map.getContainer().offsetHeight,
+        width: map.getContainer().offsetWidth,
         bounds: [
           [bbox[0], bbox[1]],
           [bbox[2], bbox[3]],
         ] as BoundsArray,
-        padding: 25,
+        breakpoint,
       }
 
       flyToBounds(map, settings)
